+++ version_metadata/__init__.py
@@ -1,0 +1,4 @@
+from .config import Config
+from .main import main
+
+
